refactor: create route group for upload and results
