+++ src/app/(slates)/layout.tsx
@@ -1,0 +1,13 @@
+import { SlateProvider } from "@/contexts";
+
+export default function SlatesRouteGroupLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <SlateProvider>
+      {children}
+    </SlateProvider>
+  );
+}
